Add files via upload
📈 SALES PREDICTION DATASET! 📈

Embark on a thrilling exploration into the dynamic world of SALES PREDICTION DATASET! This Machine Learning venture immerses us in the intricacies of sales forecasting, harnessing the power of a diverse dataset to craft a robust regression model. The project showcases a strategic implementation of algorithms, featuring the adept use of Linear Regression and the OLS method.





🧠🧠the project is focus on:-

📌Preprocessing of Data

📌Outlier treatment

📌Exploratory data analysis

📌Applying algorithms

📌Evaluation



🧠 Key Findings:

Summary of Model: A concise overview showcasing the model's strengths and capabilities.

Homogeneity: Absence of homogeneity ensures reliability across diverse scenarios.

Normality in Residuals: Residuals exhibiting normality indicate model accuracy and efficiency.

No Autocorrelation: The model demonstrates independence from temporal dependencies.

No Multicollinearity: A lack of multicollinearity ensures the stability of variable relationships.

No Overfitting and Underfitting: Striking the perfect balance for optimal model performance.







🙏 Acknowledgments:

Gratitude to luminaries such as Kumar Sundram and Mithilesh singh, whose expert guidance played a pivotal role in steering this project towards success. A heartfelt appreciation to Learnbay for facilitating enlightening sessions, empowering me with the knowledge to navigate and conquer this endeavor.

Special thanks extend to CodSoft for providing an enriching internship opportunity, contributing to the growth and application of skills in real-world scenarios.







🌐 #datascience #dataanalyst #salesprediction #machinelearning #learnbay #codsoft #internshipopportunity 🌐